Added a custom board (MaxESP3_RCX400). Added some function prototypes
Pinmaps and validations for new board MaxESP3_RCX400. Added function prototypes for some functions in Command.ino and Astro.ino. This is because Visual Micro gives compilation errors when compiling functions in an .ino file that have default values for parameters. This is fixed if the default value is in the function prototype (and a prototype exist!).